@@ -16,12 +16,6 @@
 
 #include "fix.h"
 
-#define EV_TO_KCAL_PER_MOL 14.4
-#define DANGER_ZONE 0.90
-#define MIN_CAP 50
-#define SAFE_ZONE 1.2
-#define MIN_NBRS 100
-
 namespace LAMMPS_NS {
 
 class FixQEq : public Fix {
@@ -36,6 +30,7 @@ class FixQEq : public Fix {
   void min_pre_force(int) override;
 
   double compute_scalar() override;
+  static constexpr double DANGER_ZONE = 0.90;
 
   // derived child classes must provide these functions
 
