@@ -208,14 +208,14 @@ FixDeposit::FixDeposit(LAMMPS *lmp, int narg, char **arg) :
 FixDeposit::~FixDeposit()
 {
   delete random;
-  delete [] molfrac;
-  delete [] idrigid;
-  delete [] idshake;
-  delete [] idregion;
-  delete [] vstr;
-  delete [] xstr;
-  delete [] ystr;
-  delete [] zstr;
+  delete[] molfrac;
+  delete[] idrigid;
+  delete[] idshake;
+  delete[] idregion;
+  delete[] vstr;
+  delete[] xstr;
+  delete[] ystr;
+  delete[] zstr;
   memory->destroy(coords);
   memory->destroy(imageflags);
 }
@@ -737,7 +737,7 @@ void FixDeposit::options(int narg, char **arg)
       mode = MOLECULE;
       onemols = &atom->molecules[imol];
       nmol = onemols[0]->nset;
-      delete [] molfrac;
+      delete[] molfrac;
       molfrac = new double[nmol];
       molfrac[0] = 1.0/nmol;
       for (int i = 1; i < nmol-1; i++) molfrac[i] = molfrac[i-1] + 1.0/nmol;
@@ -755,13 +755,13 @@ void FixDeposit::options(int narg, char **arg)
       iarg += nmol+1;
     } else if (strcmp(arg[iarg],"rigid") == 0) {
       if (iarg+2 > narg) error->all(FLERR,"Illegal fix deposit command");
-      delete [] idrigid;
+      delete[] idrigid;
       idrigid = utils::strdup(arg[iarg+1]);
       rigidflag = 1;
       iarg += 2;
     } else if (strcmp(arg[iarg],"shake") == 0) {
       if (iarg+2 > narg) error->all(FLERR,"Illegal fix deposit command");
-      delete [] idshake;
+      delete[] idshake;
       idshake = utils::strdup(arg[iarg+1]);
       shakeflag = 1;
       iarg += 2;
