@@ -16,6 +16,7 @@
 #include "comm.h"
 #include "error.h"
 #include "fmt/chrono.h"
+#include "tokenizer.h"
 
 #include <cstring>
 #include <ctime>
@@ -231,18 +232,25 @@ void Timer::modify_params(int narg, char **arg)
     } else if (strcmp(arg[iarg], "timeout") == 0) {
       ++iarg;
       if (iarg < narg) {
-        _timeout = utils::timespec2seconds(arg[iarg]);
-      } else
-        error->all(FLERR, "Illegal timer command");
+        try {
+          _timeout = utils::timespec2seconds(arg[iarg]);
+        } catch (TokenizerException &) {
+          error->all(FLERR, "Illegal timeout time: {}", arg[iarg]);
+        }
+      } else {
+        utils::missing_cmd_args(FLERR, "timer timeout", error);
+      }
     } else if (strcmp(arg[iarg], "every") == 0) {
       ++iarg;
       if (iarg < narg) {
         _checkfreq = utils::inumeric(FLERR, arg[iarg], false, lmp);
-        if (_checkfreq <= 0) error->all(FLERR, "Illegal timer command");
-      } else
-        error->all(FLERR, "Illegal timer command");
-    } else
-      error->all(FLERR, "Illegal timer command");
+        if (_checkfreq <= 0) error->all(FLERR, "Illegal timer every frequency: {}", arg[iarg]);
+      } else {
+        utils::missing_cmd_args(FLERR, "timer every", error);
+      }
+    } else {
+      error->all(FLERR, "Unknown timer keyword {}", arg[iarg]);
+    }
     ++iarg;
   }
 
